\name{trimmedDNAcopyAssays}
\alias{trimmedDNAcopyAssays}
\docType{data}
\title{Sample data of class 'DNAcopyAssays'}
\description{
  Encapsulates two samples from the \code{\link[DNAcopy]{coriell}} data,
  but with NA values replaced by zero.
}
\usage{data(trimmedDNAcopyAssays)}
\seealso{\code{\link[DNAcopy]{coriell}}, \code{\link{dnacopyAssays}}}
\examples{
data(trimmedDNAcopyAssays)
str(trimmedDNAcopyAssays)
}
\keyword{datasets}
